==> notebooks/modflow_setup/4_add_transport.ipynb <==
# Add transport to the MF6 flow model

import sys, os
sys.path.append('../../dependencies/')
import shutil
import flopy
import pyemu
# import specific FloPy modules
from flopy.utils.gridgen import Gridgen 
from flopy.utils.gridintersect import GridIntersect

import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt

# for working with shapefiles
import shapefile as sf
from shapely.geometry import Polygon, LineString, MultiLineString, Point

org_model_ws = os.path.join("..","..","models","sgn_mfsetup")
model_ws = org_model_ws + "_transport"
if os.path.exists(model_ws):
    shutil.rmtree(model_ws)
shutil.copytree(org_model_ws,model_ws)
sim = flopy.mf6.MFSimulation.load(sim_ws=model_ws)

model_name = "sgn_50"
m = sim.get_model(model_name)

### some trickery

Just copy the gwt model from the flopy lesson

gwt_name = "gwt-sgn"
gwf_name = "sgn_50"

gwt_org_d = os.path.join("..","..","models","sgn_model_3layer")
gwt_files = [f for f in os.listdir(gwt_org_d) if "gwt" in f.lower()]
for f in gwt_files:
    shutil.copy2(os.path.join(gwt_org_d,f),os.path.join(model_ws,f))
#shutil.copy2(os.path.join(gwt_org_d,"sim.tdis"),os.path.join(model_ws,gwf_name+"_sim.tdis"))
    
# fix tdis
lines = open(os.path.join(model_ws,gwf_name+"_sim.tdis"),'r').readlines()
with open(os.path.join(model_ws,gwf_name+"_sim.tdis"),'w') as f:
    for line in lines:
        f.write(line)
        if "begin perioddata" in line.lower():
            f.write(" 1.57680000E+09  2       1.10000000\n")
            f.write(" 1.57680000E+09  2       1.10000000\n")
            f.write("END perioddata\n")
    
# fix the dis
shutil.copy2(os.path.join(model_ws,model_name+".dis"),os.path.join(model_ws,gwt_name+".dis"))
# fix the nam file
lines = open(os.path.join(model_ws,gwt_name+".nam"),'r').readlines()
with open(os.path.join(model_ws,gwt_name+".nam"),'w') as f:
    for line in lines:
        #if "ssm" in line.lower():
        #    continue
        #if "obs" in line.lower():
        #    continue
        line = line.replace("disv","dis")
        f.write(line)
# fix the mfsim nam file
lines = open(os.path.join(model_ws,"mfsim.nam"),'r').readlines()

with open(os.path.join(model_ws,"mfsim.nam"),'w') as f:
    for line in lines:
        f.write(line)
        if "begin models" in line.lower():
            f.write(" gwt6 {0}.nam {0}\n".format(gwt_name))
        if "begin solutiongroup" in line.lower():
            f.write(" ims6 {0}.ims {0}\n".format(gwt_name))
        if "begin exchanges" in line.lower():
            f.write("GWF6-GWT6  sgn.gwfgwt  {0}  {1}\n".format(gwf_name,gwt_name))
# fix for mst so porosity is layered
lines = open(os.path.join(model_ws,gwt_name+".mst"),'r').readlines()
with open(os.path.join(model_ws,gwt_name+".mst"),'w') as f:
    for line in lines:
        if "porosity" in line:
            f.write("  porosity layered\n    CONSTANT  0.2\n    CONSTANT  0.2\n    CONSTANT 0.2\n")
            f.write("END griddata\n")
            break
        f.write(line)

ix = GridIntersect(m.modelgrid)

source_shp = os.path.join("..","..","data","sgn","shp","real_unknown_sources.shp")
shp = sf.Reader(source_shp)
print(shp.fields)

src_ij_conc = []
for s,r in zip(shp.shapes(),shp.records()):
    #print(s.points[0])
    pt = Point(*s.points[0])
    src_cell = ix.intersect(pt)
    ij_conc = list(src_cell[0][0])
    ij_conc.append(r[-2])
    if ij_conc not in src_ij_conc:
        src_ij_conc.append(ij_conc)
    #print(src_cell[0][0])

src_ij_conc

#rewrite the cnc file
cnc_file = os.path.join(model_ws,gwt_name+".cnc")
with open(cnc_file,'w') as f:
    f.write("BEGIN options\n  BOUNDNAMES\nEND options\n\n")
    f.write("BEGIN dimensions\nMAXBOUND {0}\nEND dimensions\n\n".format(len(src_ij_conc)))
    f.write("BEGIN period 1\n")
    for src_rec in src_ij_conc:
        line = "1 {0} {1} {2}".format(src_rec[0]+1,src_rec[1]+1,src_rec[2])
        if src_rec[2] == 200:
            line += " conc-1\n"
        else:
            line += " conc-2\n"
        f.write(line)
    f.write("END period 1\n")

            
    

obs_shp = os.path.join("..","..","data","sgn","shp","Target_L1_sept2019_head_PCE.shp")
shp = sf.Reader(obs_shp)
print(shp.fields)

cobs_ij = []
for s,r in zip(shp.shapes(),shp.records()):
    #print(s.points[0])
    pt = Point(*s.points[0])
    src_cell = ix.intersect(pt)
    #print(len(src_cell))
    if len(src_cell) == 0:
        print("out of grid",r[0])
        continue
    #print(src_cell)
    ij = list(src_cell[0][0])
    if ij not in cobs_ij:
        ij.append(r[0].lower().replace("_",""))
        ij.append(r[8])
        cobs_ij.append(ij)
    else:
        print("same cell",ij,r[0],[r for r in ij_conc if r == ij])

cobs_ij

with open(os.path.join(model_ws,gwt_name+".obs"),'w') as f:
    f.write("BEGIN options\n DIGITS 15\n PRINT_INPUT\nEND options\n\n")
    f.write("BEGIN continuous FILEOUT conc_obs.csv\n")
    for rec in cobs_ij:
        f.write("{0} CONCENTRATION 1 {1} {2}\n".format(rec[2],int(rec[0])+1,int(rec[1])+1))
    f.write("END continuous FILEOUT conc_obs.csv\n\n")

pyemu.os_utils.run("mf6",cwd=model_ws)

df = pd.DataFrame(cobs_ij,columns=["i","j","site","pce"])
df.to_csv(os.path.join(model_ws,"pce_obsval.csv"),index=False)

ucn = flopy.utils.HeadFile(os.path.join(model_ws,gwt_name+".ucn"),precision="double",text="concentration")

fig,axes = plt.subplots(1,2,figsize=(10,5))
times = ucn.get_times()
assert len(times) == 2
mn = 1.0e+100
mx = -1.0e+100
for t in times:
    d = ucn.get_data(totim=t)
    mn = min(mn,d.min())
    mx = max(mx,d.max())
arrs = []
for ax,t in zip(axes,times):
    d = ucn.get_data(totim=t)
    cb = ax.imshow(d[0,:,:],vmin=mn,vmax=mx)
    plt.colorbar(cb,ax=ax)
    arrs.append(d[0,:,:])
d = arrs[1] - arrs[0]
d[np.abs(d)<1.0e-3] = np.nan
plt.show()

plt.imshow(d)
plt.show()
